Работа над интеграцией микросервисов

+++ internal/pkg/user/usecase/user_uc.go
@@ -23,19 +23,19 @@ func CreateUserUsecase(userRepo user.UserRepo) *UserUsecase {
 }
 
 // GetMyProfile возвращает пользователю его профиль
-func (uc *UserUsecase) GetMyProfile(ctx context.Context, userID int) (profile *models.UserProfile, err error) {
+func (uc *UserUsecase) GetMyProfile(ctx context.Context, userID int64) (profile *models.UserProfile, err error) {
 	profile, err = uc.userRepo.GetUserProfile(ctx, userID)
 	return
 }
 
 // UpdateMyProfile обновляет профиль пользователя и возвращает обновлённый профиль
-func (uc *UserUsecase) UpdateMyProfile(ctx context.Context, userID int, data *models.UserProfileUpdateRequest) (updatedProfile *models.UserProfile, err error) {
+func (uc *UserUsecase) UpdateMyProfile(ctx context.Context, userID int64, data *models.UserProfileUpdateRequest) (updatedProfile *models.UserProfile, err error) {
 	updatedProfile, err = uc.userRepo.UpdateUserProfile(ctx, userID, *data)
 	return
 }
 
 // SetMyAvatar устанавливает пользователю аватарку
-func (uc *UserUsecase) SetMyAvatar(ctx context.Context, userID int, file *multipart.File, fileHeader *multipart.FileHeader) (updated *models.UserProfile, err error) {
+func (uc *UserUsecase) SetMyAvatar(ctx context.Context, userID int64, file *multipart.File, fileHeader *multipart.FileHeader) (updated *models.UserProfile, err error) {
 	fileName, err := uc.userRepo.SetUserAvatar(ctx, userID, uploads.ExtractFileExtension(fileHeader.Filename), int(fileHeader.Size))
 	if err != nil {
 		return nil, fmt.Errorf("SetMyAvatar: %w", err)
